fix(file): remove unuse file
