--- client/src/containers/auth/PrivateRoute.js
@@ -1,21 +1,0 @@
-import React from "react";
-import { Route, Redirect } from "react-router-dom";
-import auth from "./auth-helper";
-const PrivateRoute = ({ component: Component, ...rest }) => (
-  <Route
-    {...rest}
-    render={props =>
-      auth.isAuthenticated() ? (
-        <Component {...props} />
-      ) : (
-        <Redirect
-          to={{
-            pathname: "/signin",
-            state: { from: props.location }
-          }}
-        />
-      )
-    }
-  />
-);
-export default PrivateRoute;
